% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{add_function_hyperlinks}
\alias{add_function_hyperlinks}
\title{Add hyperlinks to function definitions}
\usage{
add_function_hyperlinks(html_files)
}
\arguments{
\item{html_files}{Character vector of file names of html files that were created
from Rmd files}
}
\description{
Finds functions that are defined in the html file(s) by looking for text of the
form \verb{foo <- function(} and then wraps \code{foo} in a \code{span} tag with \code{id="foo"}
and then whenever \code{foo} is found it wraps a \verb{a href="file.html#foo"} tag so
that it will be a hyperlink to \code{foo}'s definition.
}
\keyword{internal}