% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{add_hex_sticker}
\alias{add_hex_sticker}
\title{Add a hex sticker to package}
\usage{
add_hex_sticker(hex_png_file)
}
\arguments{
\item{hex_png_file}{The .png file with your package's hex sticker}
}
\description{
In addition to calling this function, you should add to your README.Rmd something like this:
}
\details{
\verb{# your-title <img src="man/figures/logo.png" align="right" height="139" />}

See \href{https://pkgdown.r-lib.org/reference/build_home.html#package-logo}{here}
for more.
}
