% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{add_readme}
\alias{add_readme}
\title{Add README to package}
\usage{
add_readme(rmd_file)
}
\arguments{
\item{rmd_file}{The path to a .Rmd file.}
}
\description{
This function takes a README.Rmd file, copies it into the package, and then
renders it to a README.md file.  It also adds these two files to the
.Rbuildignore.
}