% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{check_unedited}
\alias{check_unedited}
\title{Check if package directory is the unedited output of litr::render()}
\usage{
check_unedited(package_dir)
}
\arguments{
\item{package_dir}{Path to package}
}
\description{
Uses hash stored in a special \code{litr} field of DESCRIPTION file to check that
the current state of the R package directory is identical to its state at the
time that it was created by \code{litr::render()}.
}
\keyword{internal}