% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{create_from_template}
\alias{create_from_template}
\title{Internal function for creating a .Rmd file from template}
\usage{
create_from_template(pkg_name, dir, template_name)
}
\arguments{
\item{pkg_name}{Name of package to be created.}

\item{dir}{(Optional) Directory where .Rmd file should be created}

\item{template_name}{Name of template}
}
\description{
Internal function for creating a .Rmd file from template
}
\keyword{internal}