% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{description_litr_hash_field_name}
\alias{description_litr_hash_field_name}
\title{Generate litr hash field name for DESCRIPTION file}
\usage{
description_litr_hash_field_name()
}
\description{
Generate litr hash field name for DESCRIPTION file
}
\keyword{internal}
