% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{description_litr_version_field_name}
\alias{description_litr_version_field_name}
\title{Generate litr version field name for DESCRIPTION file}
\usage{
description_litr_version_field_name()
}
\description{
Generate litr version field name for DESCRIPTION file
}