% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{do_not_edit_message}
\alias{do_not_edit_message}
\title{Generate do-not-edit message to put at top of file}
\usage{
do_not_edit_message(rmd_file, type = c("R", "man", "c"))
}
\arguments{
\item{rmd_file}{Name of the Rmd file to mention}

\item{type}{Whether this is a R/ file, man/ file, or a c file}
}
\description{
Generate do-not-edit message to put at top of file
}
\keyword{internal}