% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{draft_bookdown}
\alias{draft_bookdown}
\title{Create a new litr .Rmd document for creating an R package from \code{bookdown}}
\usage{
draft_bookdown(pkg_name = "frombookdown", dir = ".")
}
\arguments{
\item{pkg_name}{Name of package to be created.}

\item{dir}{(Optional) Directory where .Rmd file should be created}
}
\description{
This template shows how to make an R package from \code{bookdown}. It creates a directory called \verb{create-[pkg_name]} in \code{dir}.  Rendering the file \code{index.Rmd}
with \code{litr::render()} creates the bookdown and package.
}
\seealso{
\code{\link{draft}} \code{\link{draft_data}} \code{\link{draft_rcpp}} \code{\link{draft_extras}} \code{\link{draft_armadillo}}
}