% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{draft_extras}
\alias{draft_extras}
\title{Create a new litr .Rmd document for creating an R package with extras}
\usage{
draft_extras(pkg_name = "withpkgdown", dir = ".")
}
\arguments{
\item{pkg_name}{Name of package to be created.}

\item{dir}{(Optional) Directory where .Rmd file should be created}
}
\description{
This creates \verb{create-[pkg_name].Rmd} that when knitted (i.e., when passed to
\code{litr::render()}) will create an R package called \code{pkg_name}.
This template shows how to make an R package that has "extras" such as a
README, vignette(s), a pkgdown site, and a hex sticker.
}
\seealso{
\code{\link{draft}} \code{\link{draft_bookdown}} \code{\link{draft_data}} \code{\link{draft_rcpp}} \code{\link{draft_armadillo}}
}