% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{draft_rcpp}
\alias{draft_rcpp}
\title{Create a new litr .Rmd document for creating an R package that uses \code{Rcpp}}
\usage{
draft_rcpp(pkg_name = "withrcpp", dir = ".")
}
\arguments{
\item{pkg_name}{Name of package to be created.}

\item{dir}{(Optional) Directory where .Rmd file should be created}
}
\description{
This creates \verb{create-[pkg_name].Rmd} that when knitted (i.e., when passed to
\code{litr::render()}) will create an R package called \code{pkg_name}.
This template shows how to make an R package that uses \code{Rcpp}.
}
\seealso{
\code{\link{draft}} \code{\link{draft_bookdown}} \code{\link{draft_data}} \code{\link{draft_extras}} \code{\link{draft_armadillo}}
}