% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{extract_label}
\alias{extract_label}
\title{Extract chunk label from the first line of an Rmd chunk.}
\usage{
extract_label(chunk_option)
}
\arguments{
\item{chunk_option}{Character vector of the first line of a chunk, i.e. ```{r chunk-name, ...}}
}
\value{
Character vector of the chunk label if it exists, empty string "" otherwise
}
\description{
Extract chunk label from the first line of an Rmd chunk.
}