% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{extract_rmd_code_chunks}
\alias{extract_rmd_code_chunks}
\title{Get a list of code chunks from an Rmd file in vector form}
\usage{
extract_rmd_code_chunks(rmd_vector)
}
\arguments{
\item{rmd_vector}{A character vector of an Rmd file that has been split by new lines}
}
\value{
A list containing information about each code chunk including the code, whether it contains references to other code chunks, and the start and end line numbers of the code chunk in the Rmd file.
}
\description{
Get a list of code chunks from an Rmd file in vector form
}