% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_package_directory}
\alias{get_package_directory}
\title{Get package directory}
\usage{
get_package_directory(package_parent_dir, package_name, input)
}
\arguments{
\item{package_parent_dir}{The directory of where the package should go (relative to the input directory)}

\item{package_name}{The name of the package}

\item{input}{The file name of the input}
}
\description{
Get package directory
}
\keyword{internal}