% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{get_params_used}
\alias{get_params_used}
\title{Get parameter values used in rendering}
\usage{
get_params_used(input, passed_params)
}
\arguments{
\item{input}{The input file to be rendered (see \code{rmarkdown::render})}

\item{passed_params}{The list of parameters that were passed to \code{render}.}
}
\description{
When the \code{params} argument of \code{rmarkdown::render()} is explicitly used, this
overrides the default that appears in \code{input}.
}
\keyword{internal}