% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{hash_package_directory}
\alias{hash_package_directory}
\title{Hash package directory}
\usage{
hash_package_directory(package_dir)
}
\arguments{
\item{package_dir}{Path to package}
}
\description{
Gets an identifier that can be used to uniquely (whp) identify the current
state of the package. This is formed by ignoring the \code{LitrId} field of the
DESCRIPTION file, which is the location where the output of this function is
stored when \code{litr::render} generates the package.
}
\keyword{internal}