% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{insert_hrefs}
\alias{insert_hrefs}
\title{Replace a function's name with a link to its definition}
\usage{
insert_hrefs(
  txt,
  function_pattern,
  where_defined,
  all_function_names,
  pkg_name,
  remove_span = FALSE
)
}
\arguments{
\item{txt}{Character vector where each element is a row of the knitted HTML file.}

\item{function_pattern}{Regular Expression passed from \code{add_function_hyperlinks} that contains all referenced functions in the document.}

\item{where_defined}{Character vector that contains the name of the file in which a function was defined.}

\item{all_function_names}{Character vector of all referenced functions in the document.}

\item{pkg_name}{Name of the package created by litr. Taken from YAML front matter.}

\item{remove_span}{Boolean argument for removing span tags. Used for minimizing code duplication.}
}
\description{
A helper function for \code{add_function_hyperlinks} that wraps references to a
function in an anchor tag with a link to the function's definition.
}
\keyword{internal}