% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{litr-package}
\alias{litr}
\alias{litr-package}
\title{Literate Programming for Writing R Packages}
\description{
Allows one to fully create an R package in a single .Rmd file.  Includes
functionality and .Rmd templates for a literate programming approach to R
package development.
}
\examples{
\dontrun{
# Make a file create-rhello.Rmd based on a template
litr::draft("rhello")
# Now call litr::render (or press Knit if in RStudio) to generate not just
# create-rhello.html, but also an R package called `rhello`.
litr::render("create-rhello.Rmd")
}
}
\seealso{
\code{\link{render}}
}
\author{
\strong{Maintainer}: Jacob Bien \email{jbien@usc.edu}

Authors:
\itemize{
  \item Jacob Bien \email{jbien@usc.edu}
  \item Patrick Vossler
}

}
\keyword{internal}