% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{litr_gitbook}
\alias{litr_gitbook}
\title{litr version of \code{bookdown::gitbook()}}
\usage{
litr_gitbook(...)
}
\arguments{
\item{...}{Parameters to be passed to \code{bookdown::gitbook()}}
}
\description{
This behaves like \code{bookdown::gitbook()} with a few differences:
\itemize{
\item It creates an R package.
\item It adds hyperlinks to function definitions whenever a function is used
elsewhere in the document.
\item It does "Knuth-style" chunk referencing with hyperlinks.
}
}