% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{litr_html_document}
\alias{litr_html_document}
\title{litr version of \code{html_document()}}
\usage{
litr_html_document(...)
}
\arguments{
\item{...}{Parameters to be passed to \code{rmarkdown::pdf_document()}}
}
\description{
This behaves like \code{rmarkdown::html_document()} with a few differences:
\itemize{
\item It creates an R package.
\item It adds hyperlinks to function definitions whenever a function is used
elsewhere in the document.
\item It does "Knuth-style" chunk referencing with hyperlinks.
}
}