% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{litrify_output_format}
\alias{litrify_output_format}
\title{Modify an existing output format to have \code{litr} behavior}
\usage{
litrify_output_format(base_format = rmarkdown::html_document)
}
\arguments{
\item{base_format}{a preexisting, non-litr output format such as \code{rmarkdown::html_document}}
}
\description{
This function modifies the \code{pre_knit()} and \code{post_processor()} functions of a
preexisting output format so that it will have the \code{litr} behavior (meaning that an R package will be created when \code{rmarkdown::render()} is called).
}
