% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{load_all}
\alias{load_all}
\title{Load complete package}
\usage{
load_all(input, output_dir = NULL, ...)
}
\arguments{
\item{input}{The input file to be rendered (see \code{rmarkdown::render})}

\item{output_dir}{By default (and in typical usage) this is NULL, meaning
that no .html/bookdown/.pdf will result.  However, when a directory is given,
the result of the litr-knitting will be saved to this location.}

\item{...}{Additional parameters to be passed to \code{devtools::load_all()}}
}
\description{
This is a litr wrapper to \code{devtools::load_all()}.  It first calls
\code{litr::render()} with \code{minimal_eval=TRUE}, then it calls
\code{devtools::load_all()} on the generated package.
}