% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_noticeable}
\alias{make_noticeable}
\title{Make error messages noticeable}
\usage{
make_noticeable(msg)
}
\arguments{
\item{msg}{Error message}
}
\description{
Since litr error messages are amid a lot of output from knitting, we'd like
the litr ones to be eye-catching.
}
\keyword{internal}