% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{modify_chunk_label}
\alias{modify_chunk_label}
\title{Modify a chunk name and add "-dup" as a suffix.
This function assumes that we are passed a chunk option line that contains a chunk label since this function is called in the context of creating a duplicate version of a referenced chunk.}
\usage{
modify_chunk_label(chunk_option)
}
\arguments{
\item{chunk_option}{Character vector of the first line of a chunk, i.e. ```{r chunk-name, ...}}
}
\value{
Character vector of the chunk option with "-dup" appended to the chunk label. i.e., ```{r chunk-name-dup, ...}
}
\description{
Modify a chunk name and add "-dup" as a suffix.
This function assumes that we are passed a chunk option line that contains a chunk label since this function is called in the context of creating a duplicate version of a referenced chunk.
}