% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{preprocess_chunk_labels}
\alias{preprocess_chunk_labels}
\title{Find an Rmd chunk label in a code chunk and modify the chunk reference delimiter before knitting}
\usage{
preprocess_chunk_labels(rmd_file)
}
\arguments{
\item{rmd_file}{A character vector of an Rmd file lines}
}
\description{
Find an Rmd chunk label in a code chunk and modify the chunk reference delimiter before knitting
}