% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{read_hash_from_description}
\alias{read_hash_from_description}
\title{Get the hash of the package from the DESCRIPTION file}
\usage{
read_hash_from_description(package_dir)
}
\arguments{
\item{package_dir}{Path to package}
}
\description{
Get the hash of the package from the DESCRIPTION file
}
\keyword{internal}
