% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{remove_rstudio_extras}
\alias{remove_rstudio_extras}
\title{Remove extra files added by usethis}
\usage{
remove_rstudio_extras(package_dir)
}
\arguments{
\item{package_dir}{Path to package}
}
\description{
Remove .Rproj and .gitignore files if they are in the package directory.
}
\keyword{internal}