% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{replace_ansi_sequences}
\alias{replace_ansi_sequences}
\title{Replace ANSI escape sequences with their HTML equivalents}
\usage{
replace_ansi_sequences(html_files)
}
\arguments{
\item{html_files}{Character vector of file names of html files that were created
from Rmd files}
}
\description{
Finds ANSI escape sequences and replaces them with HTML tags using the \code{fansi} package
}
\keyword{internal}