% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{restore_knitr_objects}
\alias{restore_knitr_objects}
\title{Return the knitr objects to their original state}
\usage{
restore_knitr_objects(original_knitr_objects)
}
\arguments{
\item{original_knitr_objects}{As returned by \code{setup()}}
}
\description{
Return the knitr objects to their original state
}
\keyword{internal}