% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{restore_state}
\alias{restore_state}
\title{Return everything \code{setup()} changed to its original state}
\usage{
restore_state(original_state)
}
\arguments{
\item{original_state}{As returned by \code{setup()}}
}
\description{
Return everything \code{setup()} changed to its original state
}
\keyword{internal}