% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{send_to_package}
\alias{send_to_package}
\title{A knitr chunk hook for writing R code and tests}
\usage{
send_to_package(before, options, envir)
}
\arguments{
\item{before}{Indicates whether this is being called before or after the
chunk code is executed}

\item{options}{Has information from the chunk}

\item{envir}{Environment}
}
\description{
This chunk hook detects whether a chunk is defining a function or dataset
to be included in the R package (looks for the \code{roxygen2} comment format \verb{#' }).
If so, then it is written to the \verb{R/} directory.  It also looks for chunks
that have one or more lines that start with \verb{test_that(} or
\verb{testthat::test_that(} (potentially with some leading whitespace).  These
chunks are then written to the \code{tests} directory of the R package.
}
\details{
When the \code{send_to} option is used, this chunk hook instead simply writes the
code chunk to the file specified.
}
\keyword{internal}