% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{setup}
\alias{setup}
\title{Code for setup chunk}
\usage{
setup(package_dir, minimal_eval)
}
\arguments{
\item{package_dir}{Directory where R package will be created}

\item{minimal_eval}{If \code{TRUE}, then only chunks with \code{usethis} commands will
be evaluated.  This can be convenient in coding when you just want to quickly
update the R package without having to wait for long evaluations to occur}
}
\description{
\itemize{
\item Creates directory where package will be. (Deletes what is currently there as
long as it appears to have been created by litr and does not have any
subsequent manual edits.)
\item Sets the root directory to this directory
\item Sets up the main chunk hook \code{litr::send_to_package()} that sends code to the
R package directory.
\item In the case that \code{minimal_eval=TRUE}, sets up an options hook for \code{eval} so
chunks are only evaluated if there is a \code{usethis} command
\item Deactivates an internal function of the \code{usethis} package
\item Redefines the document output hook to handle chunk references differently
\item Sets up a \href{https://bookdown.org/yihui/rmarkdown-cookbook/custom-engine.html}{custom language engine} called
\code{package_doc} that creates a package documentation file and then inserts
whatever the user puts in the chunk.
}
}
\details{
Returns the original state of the knitr objects that have been modified in
setup.  This allows us to return things to the previous state after we are
finished.  This is relevant in the case where litr-knitting occurs in the
current session and we don't want to leave things in a permanently modified
state.
}
\keyword{internal}