% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{write_hash_to_description}
\alias{write_hash_to_description}
\title{Write the hash of the package to the DESCRIPTION file}
\usage{
write_hash_to_description(package_dir)
}
\arguments{
\item{package_dir}{Path to package}
}
\description{
Write the hash of the package to the DESCRIPTION file
}
\keyword{internal}