% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{write_version_to_description}
\alias{write_version_to_description}
\title{Write the version of litr used to the DESCRIPTION file}
\usage{
write_version_to_description(package_dir)
}
\arguments{
\item{package_dir}{Path to package}
}
\description{
Write the version of litr used to the DESCRIPTION file
}
\keyword{internal}