% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{add_chunk_label_hyperlinks}
\alias{add_chunk_label_hyperlinks}
\title{Add hyperlinks to embedded chunks}
\usage{
add_chunk_label_hyperlinks(
  html_files,
  reference_start = "&lt;&lt;",
  reference_end = "&gt;&gt;"
)
}
\arguments{
\item{html_files}{Character vector of file names of html files that were created
from Rmd files}

\item{reference_start}{The delimiter used to indicate the start of a chunk label}

\item{reference_end}{The delimiter used to indicate the end of a chunk label}
}
\description{
Finds chunks that are referenced in the html file(s) by looking for comments
of the form \verb{###"foo"###} and then wraps \code{foo} in a \code{span} tag with \code{id="foo"}
and then whenever the chunk label \verb{<<foo>>} is found it wraps it in a
\verb{a href="file.html#foo"} tag so that it will be a hyperlink to \code{foo}'s
definition.
}
\keyword{internal}