% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{add_chunk_label_hyperlinks}
\alias{add_chunk_label_hyperlinks}
\title{Add hyperlinks to embedded chunks}
\usage{
add_chunk_label_hyperlinks(
  html_file,
  output_file = html_file,
  reference_start = "&lt;&lt;",
  reference_end = "&gt;&gt;"
)
}
\arguments{
\item{html_file}{File name of html file that was created from Rmd file}

\item{output_file}{File name to output to. Default: \code{html_file}}

\item{reference_delim}{The delimiter used to indicate a chunk label}
}
\description{
Finds chunks that are referenced in the html file by looking for comments
of the form \verb{###"foo"###} and then wraps \code{foo} in a \code{span} tag with \code{id="foo"}
and then whenever the chunk label \verb{<<foo>>} is found it wraps it in a
\verb{a href="#foo"} tag so that it will be a hyperlink to \code{foo}'s definition.
}