% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{add_function_hyperlinks}
\alias{add_function_hyperlinks}
\title{Add hyperlinks to function definitions}
\usage{
add_function_hyperlinks(html_file, output_file = html_file)
}
\arguments{
\item{html_file}{File name of html file that was created from Rmd file}

\item{output_file}{File name to output to. Default: \code{html_file}}
}
\description{
Finds functions that are defined in the html file by looking for text of the
form \verb{foo <- function(} and then wraps \code{foo} in a \code{span} tag with \code{id="foo"}
and then whenever \code{foo} is found it wraps a \verb{a href="#foo"} tag so that it be
a hyperlink to \code{foo}'s definition.
}
\keyword{internal}