% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{add_text_to_file}
\alias{add_text_to_file}
\title{Add Some Text to a File}
\usage{
add_text_to_file(
  txt,
  filename,
  location = NULL,
  req_exist = FALSE,
  pad = FALSE,
  msg = NULL
)
}
\arguments{
\item{txt}{Character vector to add to file}

\item{filename}{Name of file}

\item{location}{Specifies where text should be added. See description for more.}

\item{req_exist}{If TRUE, then throws an error if file doesn't exist}

\item{pad}{If TRUE, then when text is being added to a preexisting file, it adds a newline}

\item{msg}{An optional message to put at top of file if this is a new file}
}
\description{
The text will be added to the file at a particular line specified by
\code{location}.  The first line of \code{txt} will be on line \code{location} of the
modified file.  If \code{location} is NULL, then text is added to end of file.
If file does not exist, it is created and \code{location} is ignored (unless
\code{req_exist} is \code{TRUE}, in which case an error is thrown).
}
\keyword{internal}