% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{add_vignettes}
\alias{add_vignettes}
\title{Add one or more vignettes to package}
\usage{
add_vignettes(rmd_files)
}
\arguments{
\item{rmd_files}{A character vector of .Rmd files, each corresponding to
a vignette}
}
\description{
Add one or more vignettes to package
}