% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{find_labels}
\alias{find_labels}
\title{Find a .Rmd chunk label in a code chunk}
\usage{
find_labels(chunk_code)
}
\arguments{
\item{chunk_code}{Character vector of code from a .Rmd code chunk. Each element is a line of the code chunk.}
}
\value{
List where chunk_idx is a logical vector for each line of the chunk corresponding to whether a chunk label of the form \verb{<<label>>} was found and chunk_ids is a character vector of chunk label was found in that chunk.
}
\description{
Find a .Rmd chunk label in a code chunk
}
\keyword{internal}