% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{insert_hrefs}
\alias{insert_hrefs}
\title{Add hyperlinks to function definitions}
\usage{
insert_hrefs(
  txt,
  function_pattern,
  where_defined,
  all_function_names,
  pkg_name,
  remove_span = FALSE
)
}
\arguments{
\item{pkg_name}{Name of the package created by litr. Taken from YAML front matter}
}
\description{
Finds functions that are defined in the html file(s) by looking for text of the
form \verb{foo <- function(} and then wraps \code{foo} in a \code{span} tag with \code{id="foo"}
and then whenever \code{foo} is found it wraps a \verb{a href="file.html#foo"} tag so
that it will be a hyperlink to \code{foo}'s definition.
}
\keyword{internal}