% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{litr_html_document}
\alias{litr_html_document}
\title{litr version of \code{rmarkdown::html_document()}}
\usage{
litr_html_document(minimal_eval = FALSE, ...)
}
\arguments{
\item{minimal_eval}{If \code{TRUE}, then only chunks with \code{litr::document()} or
\code{usethis} commands will be evaluated.  This can be convenient in coding when
you just want to quickly update the R package without having to wait for long
evaluations to occur.}

\item{...}{Parameters to be passed to \code{rmarkdown::pdf_document()}}
}
\description{
This behaves like \code{rmarkdown::html_document()} with a few differences:
\itemize{
\item It creates an R package.
\item It adds hyperlinks to function definitions whenever a function is used
elsewhere in the document.
\item It does "Knuth-style" chunk referencing with hyperlinks.
}
}