% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{litr_pdf_document}
\alias{litr_pdf_document}
\title{litr version of \code{pdf_document()}}
\usage{
litr_pdf_document(...)
}
\arguments{
\item{...}{Parameters to be passed to \code{rmarkdown::pdf_document()}}
}
\description{
This behaves exactly like \code{rmarkdown::pdf_document()} except it creates an
R package.
}