% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-litr.Rmd.
\name{render}
\alias{render}
\title{Render R markdown file}
\usage{
render(input, ...)
}
\arguments{
\item{input}{The input file to be rendered (see \code{rmarkdown::render})}

\item{...}{Additional parameters to pass to \code{rmarkdown::render}}
}
\description{
Wrapper to \code{rmarkdown::render} that does some post-processing on the html
file when that is the output.  In particular, when an html file is among the
outputs, it adds hyperlinks to functions defined within the file to make it
easier for someone reading the code to see where different functions are
defined.
}