% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{with_cleanup}
\alias{with_cleanup}
\title{Add litr hash to DESCRIPTION file if error encountered}
\usage{
with_cleanup(fun, package_dir)
}
\arguments{
\item{fun}{function being called}

\item{package_dir}{directory where package is being written to}

\item{...}{arguments to be passed to \code{fun}}
}
\description{
This creates a function that calls the passed function within the context of
a try-catch.  If an error is encountered, the litr hash is still added to
the DESCRIPTION file so that future calls to \code{litr::render()} will recognize
that it can safely overwrite the package directory (i.e., no manual editing
occurred).
}
\keyword{internal}